+++ wow_vanilla_messages/src/world/v1/v12/cmsg_gmsurvey_submit.rs
@@ -45,7 +45,7 @@ impl ClientMessage for CMSG_GMSURVEY_SUBMIT {
         let survey_id = crate::util::read_u32_le(r)?;
 
         // questions: GmSurveyQuestion[10]
-        let mut questions = Vec::with_capacity(10 as usize);
+        let mut questions = Vec::with_capacity(10);
         for i in 0..10 {
             questions.push(GmSurveyQuestion::read(r)?);
         }
