+++ wow_vanilla_messages/src/world/v1/v12/msg_raid_target_update_server.rs
@@ -63,7 +63,7 @@ impl ServerMessage for MSG_RAID_TARGET_UPDATE_Server {
             }
             RaidTargetUpdateType::FULL => {
                 // raid_targets: RaidTargetUpdate[8]
-                let mut raid_targets = Vec::with_capacity(8 as usize);
+                let mut raid_targets = Vec::with_capacity(8);
                 for i in 0..8 {
                     raid_targets.push(RaidTargetUpdate::read(r)?);
                 }
